apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose==3.3.0
pytest==8.3.3
httpx==0.27.2
